# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Layal Al Zahra Salon CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** Duration in minutes */
  duration: number;
  price: number;
  isActive: boolean;
  createdAt: string;
}